Moved humaninterface to humaninput.
