+++ CommandRobotBase.java
@@ -5,8 +5,8 @@ import org.usfirst.frc4904.standard.commands.healthchecks.AbstractHealthCheck;
 import org.usfirst.frc4904.standard.commands.healthchecks.CheckHealth;
 import org.usfirst.frc4904.standard.custom.CommandSendableChooser;
 import org.usfirst.frc4904.standard.custom.TypedNamedSendableChooser;
-import org.usfirst.frc4904.standard.humaninterface.Driver;
-import org.usfirst.frc4904.standard.humaninterface.Operator;
+import org.usfirst.frc4904.standard.humaninput.Driver;
+import org.usfirst.frc4904.standard.humaninput.Operator;
 import edu.wpi.first.wpilibj.IterativeRobot;
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
