+++ humaninput/Operator.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.standard.humaninterface;
+package org.usfirst.frc4904.standard.humaninput;
 
 
 import org.usfirst.frc4904.standard.custom.Named;
@@ -7,7 +7,7 @@ import org.usfirst.frc4904.standard.custom.Named;
  * Operator specifc version of HumanInterface
  *
  */
-public abstract class Operator extends HumanInterface implements Named {
+public abstract class Operator extends HumanInput implements Named {
 	public Operator(String name) {
 		super(name);
 	}
